release (tbk v0.2 for QGIS 3.28.11) - zip file
fixed all major incompatibilities with the current QGIS LTR 3.28. This commit bundles a release for this achievement (more or less stable execution of TBk plugin and it's main algorithms within QGIS 3.28). The Plugin metadata/Provider name will temporarily be changed to "TBk for QGIS 3.28" to mark this release.

Previous commit was without zip file