feat: enable cropping coniferous rasters with extent of perimeter as alternative to masking
In TBk both coniferous rasters MG_10m & MG_10_binary are not restricted to the forested area.
By extracting them by default with the extent of the perimeter we stick to the original TBk-style / approach. If users feel the need to use the perimeter as mask, theY can by unchecking the specific option in the advanced parameters section.