feat: make forest mixture degree (coniferous raster) an optional input
TBk generation without coniferous raster resp. based only on VHM is still a goal (although currently not really functioning 2024-06-05). Thus, there's also a need for generating post processing of local densities without the input of a coniferous raster. This commit is a first try to implement this. According to TBk-development-meeting (2024-06-05) we might have to come up with a smarter (and possibly user-friendlier) solution than current one, which relies on two parameters
- mg_use: checkbox by default True
- mg_input: coniferous raster, optional, actually only optional if mg_use = False
If mg_use = True and mg_imput is not specified an error is thrown resp. the algorithm stops.
If mg_use = False it doesn't matter if mg_imput is specified or not.
This setup forces the user to make conscious decision whether to include zonal statistics deriving from a coniferous raster and the corresponding attribute (NH) from the TBK map.