feat: Merge stand maps
This algorithm merges multiple TBk stand maps to one layer and remakes the attribute ID, such that it still functions as unique identifier for stands within the return. The former attribute ID is renamed and kept as ID_stand, which is equivalent to the new ID's suffix, while the new ID's prefix corresponds to a unique value for each TBk stand map being merged. The prefix is also included as attribute ID_meta.

Squash and merge pull request #52

* feat: add plugin part / script for merging stand maps

* fix: make "Open output file after running algorithm" work

* fix: fix typos