feat: replace parameter align_to_mg by align_method
In contrast to the former version of this tool / plugin-part with parameter align_to_mg, raster alignment must not be random, if no forest mixture input with resolution 10m x 10m is at hand.

align_method is dropdown menu with three methods to choose from:

Align to origin (X,Y) = (0,0) (default): All pixel edges match coordinates = k * pixel-resolution (1.5m or 10m), where k is an integer.

Align to mixture degree raster: If forest mixture degree is among inputs and has pixel resolution 10m x 10m, all pixels of outputs VHM 10m, MG 10m and MG 10m binary are aligned to the original mixture degree, while alignment of VHM 150cm is handled as with the default method. If this method is chosen, but the preconditions for its usage are not fulfilled, the alignment is under hood switched to the default method. The advantage of aligning to the original mixture degree with resolution 10m x 10m is, that pixel of both MG 10 and MG 10m binary are not shifted.

Random / driven by extent of masks: Outputs VHM 10m, MG 10m and MG 10m are aligned to each other, but with a random offset from the origin. VHM 150cm has its own random offset.

Notes:
1) No alignment is applied to VHM detail, as this layer is only a (partial) copy of the original VHM.
2) Methods Align to origin and Align to mixture degree raster return the same outputs, if forest mixture degree (10m x 10m) is already aligned to the origin (X,Y) = (0,0). This is the case for forest mixture degree (Mishunggrad) raster layer provided by WSL with EPSG:2056.
3) Raster outputs generated with different masks and thus covering different areas, align with each other, if method chosen is either Align to origin or Align to mixture degree raster.
4) Method Random / driven by extent of masks is a legacy allowing to prepare inputs for TBk’s main algorithm Generate BK with the sole method in praxis until July 2024.