fix: inherit ID_stand’s data type from ID
previously, local densities’ attribute ID_stand was of the type integer. As this attribute is inherited from the TBk-stand-map’s attribute ID, which now can be either integer or string, this adaptation was necessary.

Note: TBk-stand-map’s attribute ID is most likely a string, if TBk-stand-map is a merge of multiple TBk-stand-maps.