refactor: handle the reading and handling of the config file using the dict_diff utility and feedbacks
The reading and handling of the config toml file has been changed in the legacy bk core algorithm but not in the modularization parts. The legacy code has been simply copied but with further modification: the code is pasted in the parent algorithm in a method that override the QgsProcessingAlgorithm prepare method. The parameters are then directly updated when calling the prepare method. The _get_input_or_config_params methods is therefore not usefull anymore