fix: remove unmaintained BK_AG from algorithms until it is fixed
Having this unmaintained algorithm in there caused an error on QGIS startup/Plugin load with TBk qgis plugin: TypeError 'str' object is not callable. The stacktrace is empty and therefore the error really hard to pinpoint down. However it is reproducible that removing the bkag_algorithm from being loaded solves the problem.