fix: apply iterative groupwise intersection of stands & local densities
Run time of intersection of stands & local densities has proven to exponentially increase with the number of involved geometries.

To overcome this time-consuming bottleneck intersection is done in iterative manner: The stands are grouped by their x_min & y_min located in grid cells (squares) having a size defined by the value the grid_width [m] (width & height). Local densities intersecting with the extent of a stand-group are selected and then feed into the intersection-algorithm with the very group of stands.

Notes:
1)	The size of the grid cells set some limit to number of geometries involved in an iteration. Although intersections with small numbers of geometries are relatively fast, there’s a tradeoff: when cells are small more geometries of local densities get selected multiple times. This inefficiency is due to an edge effect: local densities partly overlap with extents of neighboring stand-groups. So, we shouldn’t use too small cells. 9 km^2 (grid_width = 3000m) is probably a good compromise.
2)	Using the extent of stand-group rather than the stand group itself in a spatial selection is advantageous because only a single and simple geometry is involved in selecting from the potential huge number of local densities.