feat: add advanced parameter grid_cell_size
This input is used for groupwise / iterative intersection of stands and local densities, thus tackling the run time of intersection exponentially increasing with number of geometries. This parameter is experimental as the optimal cell size is unknown at the time.

If in future we kind of know the optimal cell size, we might remove this advance parameter. If so, we then must set the variable grid_with within the algorithm’s code to a fixed value [m].