fix: calculate local density metrics groupwise for stands instead of in one go
The former calculation of metric about local densities overlapping with stands done in one go a has proven to be slow for larger stand map (canton VD). Probably this is due to an issue of run time increasing exponentially with input size. Almost certain that the number of stands matters, probably also the number of local density features involved. The workaround is, to calculate metrics for max. 1000 stands per iteration. This quick fix does not really tackle the inefficient core of the metrics calculation. An alternative approach with pandas had issues of its own. And it’s unclear if that would have been more efficient / less time consuming.