fix: make export of intermediate local densities more save
after merging a list of layers into a single layer overwriting the fid with unique values is crucial to assure exportability of all features gathered in the merged layer. In this regard 'id' has proven not be reliable in all situations (so far not observed in the context of local densities). '@row_number' is a more save source for unique identifiers and thus replaces 'id'.