feat: extended Dauerwald Code with phases a/b/c (updated QGZ Template)
The TBk QGIS Project Template now contains an updated version of the Dauerwald code, that in addition to the phase now gives a secondary phase for structured stands (a/b/c) depending on the height (hdom) of the respective stand.
This also fixes theme 20 no longer being the default theme loaded/distributed (but switched back to theme 11).

Also includes auxiliary file "tbk_dauerwald_if-clause.py" that includes the syntax that is used in the template.