refactor: rename bk_hafl files and add cleared post_process.py
The included tools (simplify eliminate, vectorize highest trees, etc.) are specifically aimed at the output of bk_hafl. In the long run this may be integrated in the "create stands" routine, since it probably isn't worth much as a separate tool. For now we move it to the category C folder as integral part of bk_hafl (without it, remainders are not "complete" as they have no hdom/hmax yet.