refactor: renamed stands_clipped.gpkg to output_clipped.gpkg
These names was problematic during the fix of multiple outputs and unnecessary inputs since all previous tools in TBk_modularized used output names like output_XXXX.gpkg. This is a quick fix and a name like stands_clipped would be appropriate but the naming convention should be consistent with the other tools output.