feat: BK Generate includes hdom_diff and local densities
Added the additional steps "calculate hdom difference" and "local densities" to the BK core stand generation (with parameter flags to turn these off). This is a convenience change since most projects require these steps anyway (this change will be made obsolete/redundant to the development in branch bk_algorithm_modularization, but is small enough to be worth it).