fix: TBk WIS2 export round all tree species values and fix sum == 0
WIS.2 can't handle floats and if the sum of tree species was 100, values were copied 1:1 and not rounded (resulting in float output). This is fixed now, by rounding when reading the values (guaranteeing integer output).

Another bug was fixed: If sum of tree species was 0, default behavior now is correctly setting p100 = 100.