fix: save clipped input-VHM as .tif
Saving the VHM-input clipped by the extent of the forest mask as virtual raster layer (.vrt) is as such a fast process. But compared to a .tif the .vrt-format slows down the following preprocessing. This is very significant when clipping from larger VHM-inputs. Thus, saving the clipped VHM-input is switched from .vrt to .tif and thereby lowering overall preprocessing run time.