fix: Modularized minimal working copy after merge
This adjusts the main modularized workflow to fix issues with the tools append_attributes and integrate the cleanup.
It is a makeshift solution so far; the general handling of paths/shared parameters isn't that clear yet; also the resulting folder structure needs some fixes (.log, .config and folder tmp -> bk_process, stands_highest_trees.gpkg out of temp). Also, the creation of the project file is still lacking.